lama_inpaint.py
import os
import cv2
import numpy as np
from PIL import Image

def run_lama_inpaint(image_path):
    # Simulated inpainting logic for demo (you'll replace with real LaMa model later)
    image = cv2.imread(image_path)
    mask = np.zeros(image.shape[:2], dtype=np.uint8)
    mask[:, :20] = 255  # Simulated mask
    inpainted = cv2.inpaint(image, mask, 3, cv2.INPAINT_TELEA)
    output_path = image_path.replace(".png", "_clean.png")
    cv2.imwrite(output_path, inpainted)
    return output_path