requirements.txt
streamlit
Pillow
torch
torchvision
opencv-python
numpy